requirements.txt
numpy
torch
opencv-python
matplotlib
sklearn
pillow